appUI/requirements.txt
behave
selenium
Appium-Python-Client
tenacity